Add files via upload
In this work with Interactive Dashboard creation, I worked with the Dataset Available on the H2E website. Which is loaded up with the Sales Data and I created a Pivot table based on the Sales by Categories, Products, Stores, and a combination of Representatives of a store. 

Later on, Generated Pivot Charts based on those Pivot Tables by removing the labels, and legends for the best visuals. Followed up with the Data Labels and Chart Titles, which will add more understanding.

To this, Added the Slicers by making Connections with pivot tables to make it more interactive with all the charts. 

Finally added all these Charts to the Dashboard worksheet to make it as Presentation-Ready.